refactor result structure to separate file + black formatter
