+++ botbowl/ai/competition/__init__.py
@@ -1,2 +1,3 @@
 from .competition import *
 from .python_socket import *
+from .result_structures import *
